requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
Babel==2.11.0
carehare==1.0.2
cffi==1.15.1
channels==4.0.0
channels-rabbitmq==4.0.1
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
Django==4.1.5
django-phonenumber-field==7.0.2
django-ratelimit==4.0.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
pamqp==3.2.1
phonenumbers==8.13.4
Pillow==9.4.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymemcache==4.0.0
pyOpenSSL==23.0.0
pytz==2022.7.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
zope.interface==5.5.2
